nfnetlink: examine conntrack reply tuples as well
In some cases, the outgoing connection information might be stored in the
reply tuple, e.g. when responding to a port forward.

Extend the netlink parsing code to test the reply direction tuple as well
and swap the meaning of inbound/outbound traffic counters in this case.

Also select the proper source address based on the direction of the flow.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>