database: do not overwrite unexpectedly preexisting databases
When persistent databases are stored on an external volume, such as an
USB drive, the corresponding mount point might only become available after
nlbwmon started and initialized its memory database from scratch.

In such a case, an already existing database file might be encountered
when attempting to write the persistent database for the first time.

Avoid overwriting old data in this situation and return with an -EEXIST
error to the caller.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>